Update: NAR_language.py, NAR_language_test.py: persistency support and test cases in separate py file
